pyproject.toml
[project]
name = "miipher-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.7.0",
    "datasets>=3.6.0",
    "hydra-core>=1.3.2",
    "librosa>=0.11.0",
    "lightning>=2.5.1.post0",
    "numpy>=2.2.6",
    "peft>=0.15.2",
    "pillow>=11.2.1",
    "pyroomacoustics>=0.8.4",
    "soundfile>=0.13.1",
    "speechbrain>=1.0.0",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "tqdm>=4.67.1",
    "transformers>=4.40.0",
    "trl[peft]>=0.18.1",
    "wandb>=0.19.11",
    "webdataset>=0.2.111",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "T201",
    "COM812",
    "ISC001",
    "PGH003",
    "FBT003",
]
unfixable = [
    "F401",
    "F841",
]

pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"*.py" = [
    "D",
    "S101",
    "N802",
    "ARG",
    "S311",
    "S301",
]
"__init__.py" = [
    "F401",
]

[tool.ruff.lint.pylint]
max-args = 6


[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true

[tool.pyrefly]

#### configuring what to type check and where to import from
project_includes = ["."]
project_excludes = ["**/.[!/.]*", "**/tests"]
search_path = ["."]
import_root = ["."]
site_package_path = [".venv/lib/python3.12/site-packages"]

#### configuring your python environment
python_platform = "linux"
python_version = "3.12"
python_interpreter = ".venv/bin/python3"

#### configuring your type check settings
ignore_errors_in_generated_code = true
use_untyped_imports = true
ignore_missing_source = true

[tool.pyrefly.errors]
bad-assignment = false
invalid-argument = false
